Capitalize files
