+++ app/src/androidTest/java/com/github/swent/echo/compose/authentication/CreateProfileTest.kt
